+++ database/migrations/2025_01_16_000000_create_users_table.php
@@ -13,22 +13,24 @@ class CreateUsersTable extends Migration
      */
     public function up()
     {
-        Schema::create('users', function (Blueprint $table) {
-            $table->integer('id')->unique();
-            $table->string('name');
-            $table->string('lastName');
-            $table->string('motherLastName');
-            $table->date('dateBirth');
-            $table->integer('isMember')->unsigned();
-            $table->foreign('isMember')->references('id')->on('roles')->onDelete('cascade');
-            $table->integer('gender')->unsigned();
-            $table->foreign('gender')->references('id')->on('genders')->onDelete('cascade');
-            $table->string('email')->unique();
-            $table->timestamp('email_verified_at')->nullable();
-            $table->string('password');
-            $table->rememberToken();
-            $table->timestamps();
-        });
+        if(!Schema::hasTable('users')) {
+            Schema::create('users', function (Blueprint $table) {
+                $table->integer('id')->unique();
+                $table->string('name');
+                $table->string('lastName');
+                $table->string('motherLastName');
+                $table->date('dateBirth');
+                $table->integer('isMember')->unsigned();
+                $table->foreign('isMember')->references('id')->on('roles')->onDelete('cascade');
+                $table->integer('gender')->unsigned();
+                $table->foreign('gender')->references('id')->on('genders')->onDelete('cascade');
+                $table->string('email')->unique();
+                $table->timestamp('email_verified_at')->nullable();
+                $table->string('password');
+                $table->rememberToken();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
